Move rendering errors to a separate file
